Add content management module and MongoDB setup
Added a content management module with relevant models, repositories, request objects and services. This module supports the management of Pages, Containers, Categories and Links with validation for mandatory fields. MongoDB has been configured for both local and production environments in 'application-dev.properties' and 'application.properties' respectively. This new setup and configuration allow for dynamic content management with MongoDB as the data store.