Fix the method name in PageController
Renamed the method `getHomePage` to `deletePage` in `PageController` class. The old name was incorrect as it suggested a 'GET' operation for the homepage, however, the method was serving the 'DELETE' operation for any page, given by its 'id'. This change enhances readability and clarifies its functional role in the code.