Refactor ContainerController and services, add deletion functionality
Made several changes to the ContainerController and related service classes. Renamed "/addContainer" URL to "/add" to simplify API. Altered addContainer methods to take a single Container object, reducing complexity. A 'parentId' field is added to the Container model to clearly identify the parent of the container. Additionally, a 'deleteContainer' method was added to the Controller and services to accurately handle deletion of containers from the parent page, improving overall code maintainability and system performance.