Removed unused Request Models; Refactored Category, Page, and Container Models
The commit aims to improve code correctness and readability. Unused request models 'ReqCategory', 'ReqContainer', and 'ReqPage' are removed. The Category class has a 'parentId' field added to it to establish a clear relationship with its parent. Method signatures in the controllers and services are refactored to use the base model classes 'Category', 'Page', and 'Container', instead of their corresponding request models. ContainerService and CategoryService manage the parent field of Category. In particular, CategoryService added a 'delete' method allowing to delete a Category in a Container.