Update build version and expose port in Dockerfile
This commit includes two updates. Firstly, the version in build.gradle is incremented from 1.0.0 to 1.0.1. This mirrors the version update in the actual software. Secondly, the Dockerfile has been adjusted to expose port 8080. This allows for mapped port access from the host machine to the docker application for better networking and communication capabilities.